dct_refurb/requirements.txt
pyzmq
fastapi
uvicorn[standard]
Pillow
python-multipart